code/media-backend/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
pydantic==2.12.5
# fastapi-cors==0.0.6 (fastapi-cors is deprecated, using starlette logic in fastapi directly usually, or fastapi.middleware.cors)
python-dotenv==1.2.1
pandas
numpy
sentence-transformers
scikit-learn
scipy
openpyxl
joblib
seaborn
websockets
httpx
